QPID-2857 : Address issues found by running FindBugs against the Java codebase
git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@996393 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/java/management/tools/qpid-cli/src/org/apache/qpid/ReportGenerator.java
@@ -48,7 +48,7 @@ public class ReportGenerator implements Runnable
             generatereport();
             try
             {
-                Thread.sleep(this.interval * 60000);
+                Thread.sleep((long)this.interval * 60000L);
             }
             catch (Exception ex)
             {
